requirements.txt
google-generativeai
streamlit
pandas
os